Combine modules, store data the right way 1. Puts .dismiss command in the greeting module, not a separate one. 2. Stores data in SQLite, not JSON.
